+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `MeanBlock` algorithm
 - `ModeBlock` algorithm
+- `MedianBlock` algorithm
 ### Changed
 - Test system modified
 - `README.md` updated